Fix FetchDownloadRecipients to filter start date (#2581)
https://eaflood.atlassian.net/browse/WATER-5303

We started filtering by start date in our `FetchDownloadRecipientsService` as part of updating our returns invitation journey to start using 'due_date IS NULL' as a filter (part of the new 'dynamic due dates' functionality).

Adding `AND start_date = ?` didn't require a feature flag because it wouldn't affect the results when we are using the old filter of `due_date = ?` as well.

But we then had a last minute panic before a release because adding it in along with `AND end_date = ?` led to a massive degradation in performance. We started seeing times of 30 seconds or more in `pre-prod`, up from 1 to 2 seconds, which caused browser timeouts.

So we turned to our fr-AI-nd to figure out the issue!

> When you add `start_date >= '2024-11-01'`, PostgreSQL uses the `unique_return` index (which likely includes `start_date` and `end_date`). This changes the join order.
>
> The problem: PostgreSQL is joining to `entity_roles` BEFORE filtering on `role`, causing it to not use the `entity_roles_idx_company_id` index effectively.

Got it? 😁

It comes down to the order in which PostgreSQL joins things when the query runs. With only one `*_date = ?` condition, it filters the return logs first before joining to the other tables. Add both and it first joins the tables _then_ tries filtering by return logs.

We can control the join order by using a CTE that first returns the filtered logs. Then we join _that_ to the other tables, and voila!

However, this did impact how we applied the where clauses and bindings. With the change to use a CTE, the return log filters need only be applied once, but the licence ref filter would still need to be applied twice, and separately.

So, we tweaked the query further to then add a filtered licence document headers CTE. Now, we are only ever applying the where clauses once (though they still have to be separate).

The result is a query that is even faster than the original (less than a second)!