Start to remove batching and rework processes
Sending a notice is broken down into 3 parts.

- send the notifications to Notify
- checks their status
- update the notice

Updating the notice is covered by `UpdateEventService`. It doesn't use any batching or have interaction with Notify so we can leave it as is (though we'll rename and move it).

Checking the notifications reuses `ProcessNotificationStatusService`. Originally it was to be run once a day and check all pending notifications. This is done in batches so we'll be updating it to remove that.

We then extended it to also just do its 'thing' for a single notice. But that means we're calling a job service from a normal service, so things feel a little messy, for example, we see both their log output.

Because we're removing the batching, we can instead imagine a service that deals with just checking and recording the status for a _single_ service: `CheckNotificationStatusService`. This then becomes a more easily shareable 'service' that can be used when sending a notice, or when checking a bunch of pending notifications.

Our new `SendNoticeService` can use `CheckNotificationStatusService` directly, which means `ProcessNotificationStatusService` no longer has to worry about handling a single notice.

So, over the next set of commits we are going to work backwards from `ProcessNotificationStatusService`, removing the batching and adding `CheckNotificationStatusService`.