Make retention period configurable
The default retention period in Notify is 7 days. But WRLS has this configured as 90 days in Notify. So, we can look back at older pending notifications should we have any.

We add it as a new config value, and update the FetchNotificationsService to use it instead of a hard coded 7 days.