Update notices query's to return the message type
https://eaflood.atlassian.net/browse/WATER-5350

We have recently updated the recipients querys to handle all of the deduplication. This means the dedupe functionality is almost redundant.

This change updates the querys to return the message type. This will be either email or letter.

This change also removes the previous logic of 'both' percipient type. This is no longer needed.

The additional recipients have been updated to supply the message type.

The dedupe requires some more work to be removed / simplified completely and will be done in a later change. The only need for the 'DetermineRecipientsService' is handling additional recipients.