Update licence document header seeder (#2590)
* Update licence document header seeder

We are seeing some test failing when there are two of the same licence ref in the document header.

We know this should not be possible due to a constraint on the 'system_external_id' in 'crm.document_header'.

The test migration did not implement this unique constraint.

This change implements the unique constraint and updates the seeder to create the data in the correct way.