Drop isCurrent=true from send Rtn invites - again! (#2593)
https://eaflood.atlassian.net/browse/WATER-5364

When the Billing & Data team started checking the recipients that would be selected for the next round of return invitations, they spotted that there were some missing (when compared to a report they ran).

When they looked, they were all linked to licences that had return versions with a reason of 'Succession or transfer of licence'.

When this is the reason against the return version, the return logs we generate from it have the flag `isCurrent`, found their `metadata` field, set to 'false'.

We don't know why; it was something we just copied from the legacy code!

Another thing we copied was their query for selecting which return logs to base a returns invitation on. Their query excluded return logs where `isCurrent = false`, so we did the same.

We were told that these should be included and to remove the `WHERE` clause from our query. We made that change in [Drop rtn log isCurrent=true from send Rtn invites](https://github.com/DEFRA/water-abstraction-system/pull/2554).

Unfortunately, we inadvertently added the clause back in when some [refactoring changes](https://github.com/DEFRA/water-abstraction-system/pull/2548) to support dynamic due dates were merged in.

So, we're making the same change again!