Fix returns reminders setup 4 null dates disabled
https://eaflood.atlassian.net/browse/WATER-5303

in [Only send reminders to recipients with due date](https://github.com/DEFRA/water-abstraction-system/pull/2587) we amended `FetchReturnsRecipientsService` to only selected licences that had due returns where the due date `IS NOT NULL`.

This is part of the dynamic due dates functionality we are implementing.

Only, the feature is behind a feature flag so not yet live. Unfortunately, when this is the case it added an issue.

It was checking the notice type first and then applying the `IS NOT NULL` condition when type was reminders. It was only checking the flag for other notice types. But when it came to the bindings (the values to pass to the query) it was _only_ checking the flag. Therefore it was adding a value that the query couldn't find a match for.

This change fixes the issue.