Fix returns reminders download 4 null dates disabled (#2595)
In [Only send reminders to recipients with due date](https://github.com/DEFRA/water-abstraction-system/pull/2587), we amended `FetchDownloadRecipientsService` to only select licences that had due returns where the due date `IS NOT NULL`.

This is part of the dynamic due dates functionality we are implementing.

Only, the feature is behind a feature flag, so it's not yet live. Unfortunately, when this is the case, it causes an issue.

It was checking the notice type first and then applying the `IS NOT NULL` condition when the type was reminders. It was only checking the flag for other notice types. However, when it came to the bindings (the values to pass to the query), it was _only_ checking the flag. Therefore, it was adding a value that the query couldn't find a match for.

This change fixes the issue.