Simplify the 'DetermineRecipientsService' to only check the contact hash id
https://eaflood.atlassian.net/browse/WATER-5350

We have recently moved all the 'deduping' logic into the query's. This allows us to remove almost all the deduplication logic previously required.

The only remaining use case for removing duplicates in 'code' is when an additional recipient has been added with the same contact (same email or address). The additional recipient logic creates an exact match of the contact hash id's created by the query.
We know that the 'RecipientsService' adds the 'additionalRecipients' at the end of the recipient array. So if we 'pick' the first occurrence, we will always get the non 'single use' recipient.

This change updates the 'DetermineRecipientsService' to only return one instance of each recipient by the contact hash.

All the duplicate recipient fixtures data has been removed.