Add new SendNoticeService
This service will replace the BatchNotificationsService and be called by SubmitCheckService.

All the batching functionality is stripped out. Plus, with checking a notifications status now handled by one service, we can call it directly for each email notification, rather than kicking off the ProcessNotificationStatus job.

That also deals with an issue we spotted, where `UpdateEventService` was being called in both places, so effectively being called twice when sending a notification.

We intend this to be called directly by SubmitCheckService when all our changes are done. So, it also handles catching any errors and logging success and failure.