Simplify CreateNoticePresenter
This is the first step of moving the work being done in SubmitCheckService to create a notice _into_ CreateNoticeService.

We're going to have CreateNoticeService call the presenter, but for this it still needs the issuer. Rather than pass the whole auth object down from Submit to the Presenter, we just pass through the value we need.

Added to that, if a 'Notice' should have a type of `notification` then it is probably better that the presenter sets this.