Update notifications view to replace column (#2605)
https://eaflood.atlassian.net/browse/WATER-5230

As part of our work to send out emails to invite users to submit their returns we need to handle errors. These errors could be because of an invalid email address or an issue with either our service or GovNotify. To simplify our business process we're going to send a letter when the email fails.

In order to help with reporting and tracking of these issues we want to update the failed notification to have a link to the new paper invite notification. This PR adds a column to store that UUID.