Simplify NotificationsPresenter
Just like CreateNoticeService we're going to move the work of creating the notifications out of SubmitCheckService into CreateNotificationsService.

But, we think we can also do away with DetermineNotificationsService to simplify things yet further. This will also make CreateNotificationsService consistent with CreateNoticeService: both will take the setup data, use the presenters to format it as needed, create the record(s) and return them to SubmitCheckService.

So, our first step is to simplify the presenters. If we create the notifications one-by-one, we can both set the `createdAt` value at time of insert and kill off another service: FetchNotificationsService.

So, we started by looking to drop that property, but ended up having a good tidy up.