Drop isCurrent logic in seeder
In [Drop isCurrent=true from send Rtn invites - again!](https://github.com/DEFRA/water-abstraction-system/pull/2593) we updated the seeder to ensure one of the seeded return logs had `isCurrent` set to false so we could show that these were still being included.

Now they should be excluded, and we're about to rip through all the fetch services and ensure this is the case. The problem is they also all depend on this and the expectation that everything seeded is returned.

Going forward, we do need to ensure we seed a return log where isCurrent is false, so we can confirm they are excluded. But we're dealing with a live issue right now so this is one for the tech-debt credit card!