Update the fetch download recipients query to the improved pattern
https://eaflood.atlassian.net/browse/WATER-5369

We previously updated all the notice fetch querys to be clearer and use CTE per contact type.

This change updates the fetch download recipients query to the improved pattern.

The previous code would 'Download is the list of all possible recipients' what we instead want is to 'Download is the list of all the recipients we will send notices to.'

The tests have been updated to reflect this.