Fix issue with nil returns https://eaflood.atlassian.net/browse/WATER-5153
Following the work done in https://github.com/DEFRA/water-abstraction-system/pull/2213, `app/views/return-logs/setup/submission.njk` was updated so that the value returned when the a nil return journey had been selected changed from `nil-return` to `nilReturn`.

This caused an issue in `app/services/return-logs/setup/submit-check.service.js` when checking if the validation should be skipped for a nil return as the journey value had not been updated to take into account the change to the view.