Cover all transferred return version scenarios (#2617)
https://eaflood.atlassian.net/browse/WATER-5373

In [Fix return log isCurrent setting and handling](https://github.com/DEFRA/water-abstraction-system/pull/2612), we fixed how the `isCurrent` flag is set for return logs when a return version with the reason 'succession-or-transfer-of-licence' is added.

The PR details the history of why we set this flag, and how we had it wrong for a long time. But in short, if a licence has a return version with the reason 'succession-or-transfer-of-licence', all return logs that start before it should be flagged `isCurrent=false`, and all those that start on or after it should be flagged `isCurrent=true`.

The scenario we realise we've missed is if someone were to add a historic return version with this reason, when the current also has the reason 'succession-or-transfer-of-licence'.

For example, imagine this is the existing state.

- Return versions
  - RV1 New licence - start 2022-04-01, end 2024-03-31
  - RV2 Succeeded - start 2025-04-01

- Return logs
  - RL1 isCurrent=false - start 2022-04-01, end 2023-03-31
  - RL2 isCurrent=false - start 2023-04-01, end 2024-03-31
  - RL3 isCurrent=false - start 2024-04-01, end 2025-03-31
  - RL4 isCurrent=true - start 2025-04-01, end 2026-03-31

We then add our historic return version.

- Return versions
  - RV1 New licence - start 2022-04-01, end 2023-03-31
  - RV3 Succeeded - start 2024-04-01, end 2025-03-31
  - RV2 Succeeded - start 2025-04-01

With the code before this change, the result would be as follows.

- Return logs
  - RL1 isCurrent=false - start 2022-04-01, end 2023-03-31
  - RL2 isCurrent=false - start 2023-04-01, end 2024-03-31
  - RL3 VOID isCurrent=false - start 2024-04-01, end 2025-03-31
  - RL5 isCurrent=true - start 2024-04-01, end 2025-03-31
  - RL4 isCurrent=true - start 2025-04-01, end 2026-03-31

RL5 that has been generated _should_ have the flag set as `isCurrent=false`. This is because it is a return log that starts before RV2.

Why has this happened? The change we made in https://github.com/DEFRA/water-abstraction-system/pull/2612 was to default `isCurrent` to true when creating a return log, and then set any existing ones to false, which started before _the new return version being added_. This means that the new ones created, such as RL5, are not considered when the update runs.

This change amends the update to look for the _latest_ return version with the reason 'succession-or-transfer-of-licence' and use its start date to determine which return logs to update.