Fix previewing single use notifications in ad-hoc (#2620)
https://eaflood.atlassian.net/browse/WATER-5350

The recent change to [Simplify the 'DetermineRecipientsService' to only check the contact hash id](https://github.com/DEFRA/water-abstraction-system/pull/2603) resulted in `DetermineRecipientsService` no longer setting the contact type.

According to the comment in `AddRecipientService`, it is expecting it to do so.

```javascript
/**
 * This logic is implemented as a function to illustrate the additional data required for the download link.
 *
 * We set the 'contact_type' to 'Single use'. This will be shown in the download for the recipient. This does not
 * affect the contact type used to send the notice. This is because the 'DetermineRecipientsService' sets the
 * default 'contact_type' to 'Returns agent' when prior conditions are not met.
 *
 * We set the 'licence_ref' the same as the 'licence_refs' to allow the download to render each recipient (it fetches
 * all possible recipients without deduping).
 *
 * @private
 */
function _addDownloadRecipientData(licenceRef) {
  return {
    contact_type: 'Single use',
    licence_ref: licenceRef,
    message_type: 'Letter'
  }
}
```

Now, when you go to preview a notification for an added recipient in the ad-hoc journey, it throws an error. This is because when it hits this function...

```javascript
function _messageRef(noticeType, messageType, contactType) {
  return MESSAGE_REFS[noticeType][messageType][contactType]
}
```

...in `NotificationsPresenter`, it can find no match for contact type 'Single use', which means `messageRef` is not set, which means `pageTitle: sentenceCase(messageRef.replace(/_/g, ' ')),` in `PreviewPresenter` throws an error.

For now, we co-opt 'both' and replace it with 'Single use'. Use of 'both' was dropped when we [Updated the notice queries to return the message type](https://github.com/DEFRA/water-abstraction-system/pull/2583), along with changes to have the query de-dupe recipients.