Allow transferred rtn logs 2 be selected paper rtn
https://eaflood.atlassian.net/browse/WATER-5390

We weren't aware until the invitations were sent out for the first non-quarterly return cycle that we had misunderstood what the isCurrent flag in return logs was for and how it should be sent.

[Fix return log isCurrent setting and handling](https://github.com/DEFRA/water-abstraction-system/pull/2612), was the change which fixed our understanding of the flag (so refer to it for details).

In summary, when a new return version is added with the reason 'succeeded or transferred'.

- New return logs created on or after that return version should have isCurrent=true
- All return logs before it should have isCurrent=false

The flag is a sledgehammer approach to ensuring the new licensee cannot see the existing return logs. It's a sledgehammer approach, because setting the flag to `false` means _no one_ can see the return logs on the external site any more.

The same is therefore true for sending returns invitations and reminders: return logs flagged as `isCurrent=false` should not be included, as it might result in us sending notifications to recipients who cannot see them online any more.

The one place you do need to see them is when sending a paper return. The primary reason for the paper returns functionality is to enable Billing & Data to send paper returns to customers who can no longer view their returns on the external site.

So, this change ensures _all_ 'due' return logs are displayed on the **Select the returns for the paper return**.