Rename Fetch service to make it clearer
The FetchLetterRecipients service sounds like it would be used in cases where a recipient has an address rather than an email.

In fact, the service was written specifically for paper-returns, where all recipients get a letter, hence the use of 'letter' in the name.

We rename it to FetchPaperReturnRecipients so it is clearer in what context it is used.