Fix - missing rtn reference in ad-hoc download
Its the period start and end date we can't handle in the ad-hoc download as the user doesn't select these in the ad-hoc journey.

But there is no reason to not show the return reference, which would go some way to helping a user understand why the results seem to be duplicated.