Update CreateNotificationsService to do all work
Otherwise known as Ronseal!

Other than updating SubmitCheckService itself, this is the final step in moving the work of creating the notifications away from the submit and into the CreateNotificationsService.

We've simplified the logic around the presenters so they are used directly and consistently.

We can also use PostgreSQL's `returning()` ability to return all the information we need to pass back to SubmitCheckService. This means we can make another service, `FetchNotificationsService` redundant.