Fix failing create notifications unit test
https://eaflood.atlassian.net/browse/WATER-5119

In [Switch to non-batched simplified send notices](https://github.com/DEFRA/water-abstraction-system/pull/2602) we made a number of changes to to simplify how we send notices and their notifications to Notify.

However, we inadvertently added a unit test that asserts a value that needs to be dynamic (dependent on when the test is run) but left it as hard coded.

This fixes the broken unit test.