Fix binding order in fetch recipients services (#2638)
https://eaflood.atlassian.net/browse/WATER-5369

In [Update the fetch download recipients query to the improved pattern](https://github.com/DEFRA/water-abstraction-system/pull/2613), we updated the `FetchDownloadRecipientsService` and `FetchReturnsRecipientsService` services, not only to make them consistent with other notice queries, but also to simplify how we filter the results.

However, we made a mistake in the change. The order of the bindings is out of sync with the order in which the bindings are applied in the query. The unit tests didn't catch this, as they are based on return logs that span a single day, allowing the results to be asserted as `equal` rather than `contains`.

This change fixes the bindings to get sending returns invitations and reminders working again. We'll need to consider the unit tests another time!