Protect against floating point precision in returns
https://eaflood.atlassian.net/browse/WATER-5346

This resulted from a query by the Billing & Data team about a return with the following entries.

```csv
end date,reading,volume
2021-11-30,,
2021-12-31,,
2022-01-31,,
2022-02-28,,
2022-03-31,,
2022-04-30,,0
2022-05-31,,0
2022-06-30,,1686
2022-07-31,,2019.0000000000002
2022-08-31,,2883
2022-09-30,,1845
2022-10-31,,0

```

They queried why there was such a precise value. We know the submission was made in megalitres, which would have been converted to cubic metres. Assuming it was entered at the same precision as the others, this would be the calculation.

> 2.019 / 0.001 = 2,019

Unfortunately, we don't have a record of the _actual_ value submitted. But our best guess is that this is down to JavaScript's inability to handle decimal values natively, i.e. it is a [floating point precision error](https://en.wikipedia.org/wiki/Floating-point_error_mitigation).

Fortunately, the return log in question was submitted through the legacy code. We submitted the same values in our journey, and the result was `2,019`. But we think the difference is the versions of Node.js used then and now. Supposedly, v22 (our current version) has 'Improved decimal canonicalization', though that's AI telling us that. When the return log was submitted, we believe the legacy code was running on v16.

However, even on a later version of Node.js, we can prove that we would be susceptible to the same floating point precision errors. Here is an example from using the Node REPL.

```bash
> 2.018 / 0.001
2017.9999999999998
```

We encountered this [issue in two-part billing](https://github.com/DEFRA/water-abstraction-system/pull/1069). Our solution was to add the package [big.js](https://github.com/MikeMcl/big.js) and use it for any calculations that involve decimal values.

We've checked, and we are _not_ using **big.js** when converting submitted values to cubic metres.

In fact, we're doing the same calculation in multiple places! 🤦

So, to mitigate the possibility this change updates the return log calculations to use **big.js**, and just do the calculation in one place.