Update any pages using the updated layout to infer the 'pageHeadingHtml'
We have been updating views to 'template standard' this has involved using a shared page heading on all the updated pages to have some consistency.

However, we have not been consistent in its implementation.

This change updates all the pages using the updated 'pageContent' block to use the page heading from the layout itself.

Pages no longer need to set the 'pageHeadingHtml' on the page, this will be set in the layout, this ensures consistency.