Show exact search matches (#2573)
https://eaflood.atlassian.net/browse/WATER-5292

The new search facility just matches anything that contains the search string somewhere in the fields being searched, but that means the best matches are often not the first ones to be displayed.

This change addresses that by finding exact matches and displaying those first on the search results page.

It also includes adding a filter for the search results and UI improvements based on an updated prototype.