Update licences view page to use a layout
https://eaflood.atlassian.net/browse/WATER-5372

In preparation for moving to an updated navigation for the licence pages, we need to ensure each page renders itself (they currently share a view).

This change introduces a licence page-specific layout. This extends the base layout and adds a 'subPageContent' block.

The page-specific data will be shown in this 'subPageContent' block.