Set due date on rtn logs when confirmed as sent
https://eaflood.atlassian.net/browse/WATER-5218

We've previously made changes to our returns notice engine to capture the return log IDs linked to a recipient.

For example, when creating a returns invitation, we start with the 'due' return logs for the selected period. They give us the licence references from which we determine the recipients for the invitation.

Previously, once we had the licence references the return logs no longer mattered. Knowing this change was coming, we updated the query to return an array of the return log IDs for each recipient, populated by those that determined the licence refs, that then determined the recipient. These arrays of return log IDs are then persisted against each notification.

Because of this we can now implement this change. Once we receive confirmation that a notification has been 'sent' by Notify, we will set the due date on those return logs linked to the notification.

We'll only be doing this where the notification was sent to the licence holder or primary user (essentially letter or email) and where the notice type was 'returns invitation'.

Because of this we will never have the same return log being updated from two separate notifications. The due date we use is the one used in the original notification, which has been saved against it.