Correct changed licence refs in tear down
https://eaflood.atlassian.net/browse/WATER-5294

We did some work to try and stabilise our acceptance tests as they were frequently failing for team members. One of the changes was how data was seeded, including the licence references used for the test licences.

Previously, they had been related to expected the frequency of the returns submissions linked to the licence. But we had lots of tests that did not require return logs to be populated, so over time they just became 'generic' licence references with no relation to returns frequency.

So, the stabilisation work changed the references to `AT/TEST01` to `AT/TEST/04`. Howevere, when we made this change we overlooked updating some logic in our data teardown.