Correct changed licence refs in tear down (#2658)
https://eaflood.atlassian.net/browse/WATER-5294

We performed some work to stabilise our acceptance tests, as they were frequently failing for team members. One of the changes was how data was seeded, including the licence references used for the test licences.

Previously, they had been related to the expected frequency of the return submissions linked to the licence. But we had lots of tests that did not require return logs to be populated, so over time they just became 'generic' licence references with no relation to return frequency.

So, the stabilisation work changed the references to `AT/TEST01` to `AT/TEST/04`. However, when we made this change, we overlooked updating some logic in our data teardown.