Update licence summary to use secondary navigation
https://eaflood.atlassian.net/browse/WATER-5372

This change updates the licence pages layout. We are moving from 'tabs' to a 'secondary navigations'. This means the navigation of the page is now on the left hand side.

There was a subtle requirement to make the 'skip to main' work to go to the specific page data. This was a little complex, but we need to call it out to explain why we have spread the divs across the 'beforeContent' 'content' and 'afterContent'.

The output of this layout is we have the secondary navigation taking up one quarter and the 'main' content for the pages taking up the other three quarters.