Add dynamic due date notification columns to views
https://eaflood.atlassian.net/browse/WATER-5230
https://eaflood.atlassian.net/browse/WATER-5218

We are adding two new columns; one to notices (events) and one to notifications.

The changes to the underlying tables has been made in [Migration to enhance dynamic due date support](https://github.com/DEFRA/water-abstraction-service/pull/2741).

This change updates the views we use and in the case of notifications, ensuring the new column gets populated.

**Notices**

To notices (which is currently called `events` in both the code and the database - sorry for the confusion!) we have added a new column called `trigger_notice_id`.

As part of WATER-5230 we will record the ID of the notice containing the failed notifications that triggered the automatic notice.

**Notifications**

When we create a returns invitation notification we also capture the IDs of the return logs that determined an invitation needed sending to the recipient, as well as the due date we set in the communication.

WATER-5218 is part of the dynamic due date functionality, specifically that when we receive confirmation from Notify that the notification has been 'sent', we need to set the due date on the linked return logs.

We could just pull out `return_due_date` from the JSONB `personalisation` field. However, we're fed up of building functionality that depends on JSONB properties rather than properly defined table values. Plus, the dates are held in `DD MONTH YYYY` format, for example, `28 November 2025`.

So, when creating notifications that have a due date, we'll start recording it in this field as well as the personalisation.