Allow migrations to run in node_env test
We have a problem with removing columns from legacy migrations that are still referenced in the public migrations.

When we call `npm run migrate:test` it will wipe all existing tables from the DB. It wil then run the legacy migrations to recreate them, then the public migrations.

But because we are simply trying to get to the 'current' state of the tables with the legacy migrations, if a public view migration references a deleted column we no longer create in the legacy migration it fails.

We don't want to keep columns in our legacy migrations that we've since deleted in the 'real' tables. But we also cannot delete the public migrations that added them. If we do so we'll corrupt the 'down' migration process.

So, we do like this change. We remove the reference from the migrations that refer to the delted column. It might result in a migration that now essentially does nothing. But as the column no longer exists we don't really have a need or desire to get back to a state where it does.