Add sub navigation to licence summary secondary navigation
https://eaflood.atlassian.net/browse/WATER-5372

This change updates the licence pages layout. We are moving from 'tabs' to a 'secondary navigations'. This means the navigation of the page is now on the left hand side.

The output of this layout is we have the secondary navigation taking up one quarter and the 'main' content for the pages taking up the other three quarters.

This change adds the licence summary sub navigation links. These are only shown on the licence summary page.