Break out check status unit tests by notice type
We're introducing something that returns invitations do that returns reminders don't.

So, it makes it more awkward to try and cover this, when we are just using returns invite to check a letter, and returns reminder to check an email.

We'd already added abstraction alerts separately, so using that as a starting point we break up the unit tests by notice type, then method.

Where a notice type can be sent as both letter and email, such as abstraction alerts, we ensure both are covered. We also add paper return to the tests, having realised they are the only notice type that can expect to get a cancelled status returned from Notify. It also ensure all known notice types are covered.

This makes it easier to then check our changes for returns invitations and setting the due date.