Ensure CreateNotifications returns dueDate
We were already returning returnLogIds but only due date is it has been set from the relevant presenter.

By adding it into the `.returning()` call we ensure the result is consistent. Plus, the value we get back is the one from the DB which means it is an actual date, as opposed to a string representation.